finished donate page, updated page-user styles, updated navigation links and pages
